simplify token generation
