+++ gen-tokens/package.json
@@ -2,27 +2,14 @@
   "name": "gen-tokens",
   "version": "1.0.0",
   "description": "",
+  "type": "module",
   "main": "index.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
-    "dev": "http-server ./dist",
-    "build": "webpack"
+    "test": "echo \"Error: no test specified\" && exit 1"
   },
   "author": "",
   "license": "ISC",
   "dependencies": {
-    "@babel/core": "^7.16.0",
-    "babel-loader": "^8.2.3",
-    "babel-preset-es2015": "^6.24.1",
-    "firebase": "^9.2.0",
-    "webpack": "^5.61.0",
-    "webpack-dev-server": "^4.4.0"
-  },
-  "devDependencies": {
-    "clean-webpack-plugin": "^4.0.0",
-    "copy-webpack-plugin": "^9.0.1",
-    "html-webpack-plugin": "^5.5.0",
-    "http-server": "^14.0.0",
-    "webpack-cli": "^4.9.1"
+    "firebase": "^9.2.0"
   }
 }
